app_manager/app/requirements.txt
psycopg2-binary~=2.9.10
pydantic~=2.11.7
pydantic-settings~=2.10.1
pandas~=2.3.1
fastapi~=0.116.1
openpyxl~=3.1.5